chuc_nang_cua_hieu

+++ app/src/main/java/hieunnph32561/fpoly/du_an_1_hieu/dao/hoadonDAO.java
@@ -39,7 +39,7 @@ public class hoadonDAO {
                 @SuppressLint("Range") HoaDon s = new HoaDon(
                         cursor.getInt(cursor.getColumnIndex("maHD")),
                         cursor.getInt(cursor.getColumnIndex("maTk")),
-                        (int) cursor.getDouble(cursor.getColumnIndex("tongTien")),
+                        (double) cursor.getDouble(cursor.getColumnIndex("tongTien")),
                         ngay,
                         cursor.getInt(cursor.getColumnIndex("trangThai")),
                         cursor.getString(cursor.getColumnIndex("phuongThuc"))
